renamed links bar
